aistock_assistant_final/requirements.txt
# A股智能助手 - Python依赖列表
# 使用方法: pip install -r requirements.txt

# GUI框架
PyQt5>=5.15.0

# 数据处理
pandas>=1.3.0
numpy>=1.20.0
numexpr>=2.7.3

# 数据源
pytdx>=1.72
akshare>=1.10.0

# 任务调度
apscheduler>=3.9.0

# HTTP请求
requests>=2.26.0

# 配置文件
pyyaml>=5.4.0

# 日志
colorama>=0.4.4

# 数据库
sqlalchemy>=1.4.0

# 图表
matplotlib>=3.4.0

# 科学计算
scipy>=1.7.0
scikit-learn>=0.24.0

# 文档处理
python-docx>=0.8.11
PyPDF2>=2.0.0

# 图像处理
Pillow>=8.0.0

# 其他
lxml>=4.6.0
